requirements.txt
djangorestframework
django
pillow